backend-lti/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
PyJWT[crypto]==2.8.0
cryptography==42.0.0
redis==5.0.1
python-multipart==0.0.6
httpx==0.27.0
requests==2.31.0
